requirements.txt
py-cord==2.5.0
pytz==2024.1
python-decouple==3.8
requests==2.31.0
lyricsgenius==3.0.1
aiohttp==3.9.3
PyGithub==2.3.0
geopy==2.4.1
timezonefinder==6.5.0
watchdog==4.0.0
openai==1.12.0
psutil==5.9.8
schedule==1.2.1
PyYAML==6.0.1
aiosignal==1.3.1
aiosqlite==0.19.0
async-timeout==4.0.2
attrs==23.1.0
charset-normalizer==3.1.0
colorama==0.4.6
frozenlist==1.4.1
idna==3.4
multidict==6.0.4
prettytable==3.8.0
wcwidth==0.2.6
yarl==1.9.2
humanize==4.9.0
